Refactored Homepage and Products index pages
